Add checked-in build definitions
These build definitions will be used by pipebuild to build the product.

The signing now uses MicroBuild because the builds are being moved to
the DevDiv VSTS instance, and the Microsoft Code Signing system which
was used in MSEng is not supported there.

MicroBuild does not support the .NET Core MSBuild, so the full-framework
MSBuild is used to build the signing targets.  However, the new build
machines do not have MSBuild 15, so they use 14. This means that
projects which use the new MSBuild 15 syntax fail to be imported.
They're not needed for the signing, so their importing is skipped by
using a new SkipMsBuild15Projects property.